requirements.txt
torch>=1.11.0
torchvision
numpy>=1.21.5
scikit-image>=0.19.2
pillow
einops
